recai/recai-back/requirements.txt
fastapi
uvicorn
pydantic-settings
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
pytest-cov>=4.1.0
ruff>=0.4   